Update install instructions for fedora
There is a Copr repository available for Fedora 35 - Rawhide and OpenSuse tumbleweed, however the repository is lacking the Leap branch of OpenSuse. It should be placed into the OBS or games:tools repository some time in the near future. CentOS and RHEL don't have any packages available (to my current knowledge). As you have to manually download the .repo file in OpenSuse, OpenSuse instructions won't be included in this pr.

This was created to save new people (yk flooding in) from not being able to install Prism Launcher without having to surf through the discord server. 

Signed-off-by: Ascende Supirius <92772508+LackingCats@users.noreply.github.com>