Update cf-compromised-alert.md to use more descriptive language
Changes cf-compromised-alert.md to indicate that you have to download the Windows script to your user's downloads folder and that the Linux script downloads before running. Should be a minor change that helps the scripts be more accessible for people who haven't run many similar things before. I don't know if this will break formatting due to the new text being potentially too long; this probably needs to be checked.

Signed-off-by: Raamesh <raameshb@proton.me>